refactor(escrow): move escrow types to a new file for better organization
