--- apps/web/lib/types/escrow.types.ts
@@ -1,92 +1,0 @@
-import { CreatedAt, UpdatedAt } from "./date.types";
-import { EscrowEndpoint, HttpMethod, Status } from "./utils.types";
-
-export type MilestoneStatus = "completed" | "approved" | "pending";
-
-export type Milestone = {
-  description: string;
-  status?: MilestoneStatus;
-  flag?: boolean;
-};
-export interface Escrow {
-  id: string;
-  title: string;
-  description: string;
-  createdAt: CreatedAt;
-  updatedAt: UpdatedAt;
-  contractId?: string;
-  balance?: string;
-  token: string;
-  milestones: Milestone[];
-  serviceProvider: string;
-  engagementId: string;
-  disputeResolver: string;
-  amount: string;
-  platformAddress: string;
-  platformFee: string;
-  approver: string;
-  releaseSigner: string;
-  user: string;
-  issuer: string;
-  disputeFlag?: boolean;
-  releaseFlag?: boolean;
-  resolvedFlag?: boolean;
-}
-
-// Escrow's Payload
-export type EscrowPayload = Omit<
-  Escrow,
-  "user" | "createdAt" | "updatedAt" | "id"
->;
-
-export type ChangeMilestoneStatusPayload = {
-  contractId?: string;
-  milestoneIndex: string;
-  newStatus: MilestoneStatus;
-  serviceProvider?: string;
-};
-
-export type ChangeMilestoneFlagPayload = Omit<
-  ChangeMilestoneStatusPayload,
-  "serviceProvider" | "newStatus"
-> & {
-  approver?: string;
-  newFlag: boolean;
-};
-
-export type StartDisputePayload = Pick<Escrow, "contractId"> & {
-  signer: string;
-};
-
-export type ResolveDisputePayload = Pick<Escrow, "contractId"> &
-  Partial<Pick<Escrow, "disputeResolver">> & {
-    approverFunds: string;
-    serviceProviderFunds: string;
-  };
-
-export type EditMilestonesPayload = {
-  contractId: string;
-  escrow: EscrowPayload;
-  signer: string;
-};
-
-// Escrow's Response
-export type EscrowRequestResponse = {
-  status: Status;
-  unsignedTransaction?: string;
-};
-
-export type InitializeEscrowResponse = {
-  contract_id: string;
-  escrow: EscrowPayload;
-  message: string;
-  status: Status;
-};
-
-export type FundEscrowResponse = {
-  message: string;
-  status: Status;
-};
-
-export type SendTransactionResponse = InitializeEscrowResponse &
-  FundEscrowResponse;
